improve isolated node test
