+++ crates/core/tests/isolated_node_regression.rs
@@ -1,0 +1,240 @@
+//! Regression tests for isolated node functionality
+//!
+//! These tests ensure that isolated nodes (nodes with no peer connections) operate correctly:
+//! 1. PUT operations cache contracts locally without network timeouts (PR #1781)
+//! 2. GET operations retrieve from local cache without self-routing attempts (PR #1806)
+//! 3. Complete PUT→GET workflow functions properly on isolated nodes
+
+use freenet::{
+    config::{ConfigArgs, NetworkArgs, SecretArgs, WebsocketApiArgs},
+    dev_tool::TransportKeypair,
+    local_node::NodeConfig,
+    server::serve_gateway,
+    test_utils::{load_contract, make_get, make_put, verify_contract_exists},
+};
+use freenet_stdlib::{
+    client_api::{ClientRequest, ContractResponse, HostResponse, WebApi},
+    prelude::*,
+};
+use futures::FutureExt;
+use std::{
+    net::Ipv4Addr,
+    sync::{LazyLock, Mutex},
+    time::Duration,
+};
+use tokio::{select, time::timeout};
+use tokio_tungstenite::connect_async;
+use tracing::error;
+
+static RNG: LazyLock<Mutex<rand::rngs::StdRng>> = LazyLock::new(|| {
+    use rand::SeedableRng;
+    Mutex::new(rand::rngs::StdRng::from_seed(
+        *b"regression_test_seed_01234567890",
+    ))
+});
+
+/// Helper to create a node configuration for testing
+async fn create_test_node_config(
+    is_gateway: bool,
+    ws_api_port: u16,
+    network_port: Option<u16>,
+) -> anyhow::Result<(ConfigArgs, tempfile::TempDir)> {
+    let temp_dir = tempfile::tempdir()?;
+    let key = TransportKeypair::new();
+    let transport_keypair = temp_dir.path().join("private.pem");
+    key.save(&transport_keypair)?;
+    key.public().save(temp_dir.path().join("public.pem"))?;
+
+    let config = ConfigArgs {
+        ws_api: WebsocketApiArgs {
+            address: Some(Ipv4Addr::LOCALHOST.into()),
+            ws_api_port: Some(ws_api_port),
+        },
+        network_api: NetworkArgs {
+            public_address: Some(Ipv4Addr::LOCALHOST.into()),
+            public_port: network_port,
+            is_gateway,
+            skip_load_from_network: true,
+            gateways: Some(vec![]), // Empty gateways for isolated node
+            location: Some({
+                use rand::Rng;
+                RNG.lock().unwrap().random()
+            }),
+            ignore_protocol_checking: true,
+            address: Some(Ipv4Addr::LOCALHOST.into()),
+            network_port,
+            bandwidth_limit: None,
+            blocked_addresses: None,
+        },
+        config_paths: freenet::config::ConfigPathsArgs {
+            config_dir: Some(temp_dir.path().to_path_buf()),
+            data_dir: Some(temp_dir.path().to_path_buf()),
+        },
+        secrets: SecretArgs {
+            transport_keypair: Some(transport_keypair),
+            ..Default::default()
+        },
+        ..Default::default()
+    };
+
+    Ok((config, temp_dir))
+}
+
+/// Test complete PUT-then-GET workflow on isolated node
+///
+/// This regression test verifies isolated nodes operate correctly:
+/// - PUT operations cache contracts locally without network timeouts (PR #1781)
+/// - GET operations retrieve from local cache without self-routing attempts (PR #1806)
+/// - Complete workflow functions properly without peer connections
+#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
+async fn test_isolated_node_put_get_workflow() -> anyhow::Result<()> {
+    freenet::config::set_logger(Some(tracing::level_filters::LevelFilter::INFO), None);
+
+    // Start a single isolated node (no peers)
+    let ws_port = 50700;
+    let network_port = 50701;
+    let (config, temp_dir) = create_test_node_config(true, ws_port, Some(network_port)).await?;
+
+    // Load test contract and state
+    const TEST_CONTRACT: &str = "test-contract-integration";
+    let contract = load_contract(TEST_CONTRACT, vec![].into())?;
+    let contract_key = contract.key();
+    let initial_state = freenet::test_utils::create_empty_todo_list();
+    let wrapped_state = WrappedState::from(initial_state);
+
+    // Start the node
+    let node_handle = {
+        let config = config.clone();
+        async move {
+            let built_config = config.build().await?;
+            let node = NodeConfig::new(built_config.clone())
+                .await?
+                .build(serve_gateway(built_config.ws_api).await)
+                .await?;
+            node.run().await
+        }
+        .boxed_local()
+    };
+
+    // Run the test with timeout
+    let test_result = timeout(Duration::from_secs(60), async {
+        // Give node time to start - critical for proper initialization
+        println!("Waiting for node to start up...");
+        tokio::time::sleep(Duration::from_secs(10)).await;
+        println!("Node should be ready, proceeding with test...");
+
+        // Connect to the node
+        let url = format!(
+            "ws://localhost:{}/v1/contract/command?encodingProtocol=native",
+            ws_port
+        );
+        let (ws_stream, _) = connect_async(&url).await?;
+        let mut client = WebApi::start(ws_stream);
+
+        println!("Step 1: Performing PUT operation to cache contract locally");
+
+        // Perform PUT operation - this should cache the contract locally
+        let put_start = std::time::Instant::now();
+        make_put(&mut client, wrapped_state.clone(), contract.clone(), false).await?;
+
+        // Wait for PUT response
+        let put_result = timeout(Duration::from_secs(30), client.recv()).await;
+        let put_elapsed = put_start.elapsed();
+
+        match put_result {
+            Ok(Ok(HostResponse::ContractResponse(ContractResponse::PutResponse { key }))) => {
+                assert_eq!(key, contract_key);
+                println!("PUT operation successful in {:?}", put_elapsed);
+            }
+            Ok(Ok(other)) => {
+                panic!("Unexpected PUT response: {:?}", other);
+            }
+            Ok(Err(e)) => {
+                panic!("PUT operation failed: {}", e);
+            }
+            Err(_) => {
+                panic!("PUT operation timed out");
+            }
+        }
+
+        // Verify contract was cached locally (PR #1781)
+        assert!(
+            verify_contract_exists(temp_dir.path(), contract_key).await?,
+            "Contract should be cached locally after PUT"
+        );
+        println!("Contract verified in local cache");
+
+        println!("Step 2: Performing GET operation using local cache");
+
+        // Now perform GET operation - should use local cache without self-routing
+        let get_start = std::time::Instant::now();
+        make_get(&mut client, contract_key, true, false).await?;
+
+        // Wait for GET response
+        let get_result = timeout(Duration::from_secs(10), client.recv()).await;
+        let get_elapsed = get_start.elapsed();
+
+        // REGRESSION TEST: Verify GET completed quickly using local cache (PR #1806)
+        assert!(
+            get_elapsed < Duration::from_secs(5),
+            "GET from local cache should be fast, not hanging on self-routing. Elapsed: {:?}",
+            get_elapsed
+        );
+
+        // Verify GET succeeded with correct data
+        match get_result {
+            Ok(Ok(HostResponse::ContractResponse(ContractResponse::GetResponse {
+                contract: recv_contract,
+                state: recv_state,
+                ..
+            }))) => {
+                assert_eq!(
+                    recv_contract
+                        .as_ref()
+                        .expect("Contract should be present")
+                        .key(),
+                    contract_key
+                );
+                assert_eq!(recv_state, wrapped_state);
+                println!(
+                    "GET operation successful from local cache in {:?}",
+                    get_elapsed
+                );
+            }
+            Ok(Ok(other)) => {
+                panic!("Unexpected GET response: {:?}", other);
+            }
+            Ok(Err(e)) => {
+                panic!("GET operation failed: {}", e);
+            }
+            Err(_) => {
+                panic!("GET operation timed out");
+            }
+        }
+
+        println!("PUT-then-GET workflow completed successfully without self-routing");
+
+        // Properly close the client
+        client
+            .send(ClientRequest::Disconnect { cause: None })
+            .await?;
+        tokio::time::sleep(Duration::from_millis(100)).await;
+
+        Ok::<(), anyhow::Error>(())
+    });
+
+    // Run node and test concurrently
+    select! {
+        _ = node_handle => {
+            error!("Node exited unexpectedly");
+            panic!("Node should not exit during test");
+        }
+        result = test_result => {
+            result??;
+            // Give time for cleanup
+            tokio::time::sleep(Duration::from_secs(1)).await;
+        }
+    }
+
+    Ok(())
+}
